[test][ci] Split AA DVRT CDC version-swap tests and rename for clarity
The monolithic TestAaVersionSwapRecordTransformer (~13 min runtime across
7 tests) exceeds the 15-min per-shard CI budget even when given its own
shard. Split by topic into two classes that each fit comfortably, and
rename to better reflect what's actually under test (the DVRT-based CDC
consumer's behavior under AA version swaps, not the RecordTransformer
class itself).

- AbstractDvrtCdcAaVersionSwapTest — shared multi-region cluster setup,
  helpers (createAaHybridStore, emptyPush, writeRecords, buildClientConfig,
  pollUntilRangeObserved, drainForDuration, assertNoLoss, deleteStoreQuietly),
  and constants. Pattern matches AbstractTestRepush.
- TestDvrtCdcAaVersionSwap — single-swap basic + edge scenarios: stateful,
  stateless, mid-swap restart, watchdog timeout, buffer pressure (~440s).
- TestDvrtCdcAaVersionSwapMultiVersion — multi-version scenarios:
  back-to-back swaps, rollback (~387s).

Each subclass lands in its own dedicated shard (86 and 87).
